refactor: extract open identity hooks
